Add month pagination to trades, YTD P&L on dashboard, fractional shares, fix LEAPS form
- Add month-based pagination with prev/next navigation to trade history
- Show YTD trade income/losses/net P&L cards on Dashboard
- Allow fractional shares (float) in portfolio holdings
- Fix LEAPS form defaulting option_type to call when switching holding type

+++ IncomePilot/backend/app/models/holding.py
@@ -15,7 +15,7 @@ class Holding(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     symbol: Mapped[str] = mapped_column(String(10), nullable=False, index=True)
-    shares: Mapped[int] = mapped_column(Integer, nullable=False)
+    shares: Mapped[float] = mapped_column(Float, nullable=False)
     avg_cost: Mapped[float] = mapped_column(Float, nullable=False)
     owner: Mapped[str] = mapped_column(
         String(20), nullable=False, default="Venky"
